create ControllerUtils and move related functions there
